+++ lib/src/core/domain/place.dart
@@ -1,0 +1,14 @@
+/// Модель заведения
+class Place {
+  final int id;
+  final String name;
+  final String description;
+  final String imageUrl;
+
+  const Place({
+    required this.id,
+    required this.name,
+    required this.description,
+    required this.imageUrl,
+  });
+}