+++ lib/src/features/main/main_screen.dart
@@ -1,60 +1,81 @@
 import 'package:flutter/material.dart';
-import 'package:flutter_gen/gen_l10n/app_localizations.dart';
 import 'package:flutter_mobx/flutter_mobx.dart';
-import 'main_screen_vm.dart';
+import 'package:where_to_go_today/src/features/main/main_vm.dart';
+import 'package:where_to_go_today/src/localization/l10n.dart';
 
-/// Глвный экран с табами
-class MainScreen extends StatelessWidget {
-  static const double _labelSize = 12.0;
+/// Главный экран с табами
+class MainScreen extends StatefulWidget {
+  final MainVm vm;
+  final String? initialScreen;
+  final List<Widget> screens;
 
-  final MainScreenVm vm;
+  const MainScreen({
+    Key? key,
+    required this.vm,
+    required this.initialScreen,
+    required this.screens,
+  }) : super(key: key);
 
-  // TODO(artem-zaitsev): здесь будут храниться экраны, когда появятся
-  final List<Widget> _screens = [
-    const Center(
-      child: Text('Screen 1'),
-    ),
-    const Center(
-      child: Text('Screen 1'),
-    ),
-    const Center(
-      child: Text('Screen 2'),
-    ),
-  ];
+  @override
+  State<MainScreen> createState() => _MainScreenState();
+}
 
-  MainScreen({Key? key, required this.vm}) : super(key: key);
+class _MainScreenState extends State<MainScreen>
+    with SingleTickerProviderStateMixin {
+  late final TabController _tabController;
+
+  @override
+  void initState() {
+    super.initState();
+    final initialTabId = int.tryParse(widget.initialScreen ?? '');
+    _tabController = TabController(
+      initialIndex: initialTabId ?? 0,
+      length: widget.screens.length,
+      animationDuration: Duration.zero,
+      vsync: this,
+    );
+  }
+
+  @override
+  void dispose() {
+    _tabController.dispose();
+    super.dispose();
+  }
 
   @override
   Widget build(BuildContext context) {
     return Observer(
       builder: (_) => Scaffold(
-        body: _screens[vm.currentIndex],
+        body: TabBarView(
+          controller: _tabController,
+          physics: const NeverScrollableScrollPhysics(),
+          children: widget.screens,
+        ),
         bottomNavigationBar: BottomNavigationBar(
-          onTap: vm.tabClicked,
-          currentIndex: vm.currentIndex,
-          selectedFontSize: _labelSize,
+          onTap: _tabClicked,
+          currentIndex: _tabController.index,
           items: [
-            // TODO(any): поменять иконки из дизайна, когда они появятся
             BottomNavigationBarItem(
               icon: const Icon(Icons.search),
-              activeIcon: const Icon(
-                Icons.search,
-              ),
-              label: AppLocalizations.of(context)!.placesTabName,
+              label: context.l10n.placesTabName,
             ),
             BottomNavigationBarItem(
               icon: const Icon(Icons.card_giftcard),
-              activeIcon: const Icon(Icons.card_giftcard),
-              label: AppLocalizations.of(context)!.cardTabName,
+              label: context.l10n.cardTabName,
             ),
             BottomNavigationBarItem(
               icon: const Icon(Icons.account_circle),
-              activeIcon: const Icon(Icons.account_circle),
-              label: AppLocalizations.of(context)!.profileTabName,
+              label: context.l10n.profileTabName,
             ),
           ],
         ),
       ),
     );
+  }
+
+  void _tabClicked(int index) {
+    setState(() {
+      _tabController.index = index;
+    });
   }
 }